Intermediate commit, added executable for 3point contractions and derivative operator, but not cross-checked yet.

+++ src/L0/Core/Propagator/Propagator_shift.cpp
@@ -2,7 +2,7 @@
 
 namespace Core
 {
-  Propagator &Propagator::shift(Base::SpaceTimeIndex const idx, Base::Direction const dir, int const times)
+  Propagator &Propagator::shift(Base::SpaceTimeIndex const idx, Base::Direction const dir, size_t const times)
   {
     if (times == 0)
       return *this;
